server/pyproject.toml
[project]
name = "task-list-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-serializer>=1.4.12",
]
